requirements.txt
colorama
coloredlogs
humanfriendly
lxml
qrcode
requests
six
requests_toolbelt
pyqrcode
certifi
argparse
sqlcipher3-binary
